showWorkspaceBuildSettings
